+++ rspotify-http/Cargo.toml
@@ -11,15 +11,17 @@ keywords = ["spotify", "api"]
 edition = "2018"
 
 [dependencies]
-async-stream = { version = "0.3.0", optional = true }
+rspotify-model = { path = "../rspotify-model", version = "0.10.0" }
+
 # Temporary until https://github.com/rust-lang/rfcs/issues/2739, for
 # `maybe_async`.
 async-trait = { version = "0.1.48", optional = true }
 base64 = "0.13.0"
 futures = { version = "0.3.8", optional = true }
 log = "0.4.11"
-maybe-async = "0.2.1"
+maybe-async = "0.2.4"
 reqwest = { version = "0.11.0", default-features = false, features = ["json", "socks"], optional = true }
+serde_json = "1.0.57"
 thiserror = "1.0.20"
 ureq = { version = "2.0", default-features = false, features = ["json", "cookies"], optional = true }
 url = "2.1.1"
@@ -45,7 +47,7 @@ reqwest-native-tls-vendored = ["reqwest/native-tls-vendored"]
 ureq-rustls-tls = ["ureq/tls"]
 
 # Internal features for checking async or sync compilation
-__async = ["async-trait", "async-stream", "futures"]
+__async = ["async-trait", "futures"]
 __sync = ["maybe-async/is_sync"]
 
 [package.metadata.docs.rs]
