+++ src/endpoints/oauth.rs
@@ -1,0 +1,10 @@
+use crate::rspotify::{prelude::*, OAuth};
+
+pub trait OAuthClient: BaseClient {
+    fn get_oauth(&self) -> &OAuth;
+
+    fn user_endpoint(&self) {
+        println!("Performing OAuth request");
+        self.endpoint_request();
+    }
+}
